reorganizte utility functions in parquet and filasystem modules; support pathlib.Path in Filesystem API
